+++ crane.h
@@ -9,11 +9,11 @@
 #ifndef crane_h
 #define crane_h
 
-void getApplicationLogFile(char * buf, int maxl, const char* appName);
-void getApplicationDataDirectory(char * buf, int maxl, const char* appName);
-void getApplicationScriptsDirectory(char * buf, int maxl, const char* appName);
-void getUserHomeDirectory(char * buf, int maxl);
-void getTemporaryDirectory(char * buf, int maxl);
+void getApplicationLogFile(char * buf, size_t maxl, const char* appName);
+void getApplicationDataDirectory(char * buf, size_t maxl, const char* appName);
+void getApplicationScriptsDirectory(char * buf, size_t maxl, const char* appName);
+void getUserHomeDirectory(char * buf, size_t maxl);
+void getTemporaryDirectory(char * buf, size_t maxl);
 
 
 #endif /* crane_h */
